chore(typescript-estree): remove `createError`
