+++ packages/typescript-estree/src/check-modifiers.ts
@@ -2,12 +2,12 @@ import * as ts from 'typescript';
 
 import type { TSNode } from './ts-estree';
 
+import { TSError } from './errors';
 import { getDecorators, getModifiers } from './getModifiers';
 import {
   getDeclarationKind,
   hasModifier,
   isThisIdentifier,
-  createError,
 } from './node-utils';
 
 const SyntaxKind = ts.SyntaxKind;
@@ -131,7 +131,7 @@ function nodeHasIllegalDecorators(
 export function checkModifiers(node: ts.Node): void {
   // typescript<5.0.0
   if (nodeHasIllegalDecorators(node)) {
-    throw createError(
+    throw new TSError(
       node.illegalDecorators[0],
       'Decorators are not valid here.',
     );
@@ -144,12 +144,12 @@ export function checkModifiers(node: ts.Node): void {
     // `checkGrammarModifiers` function in typescript
     if (!nodeCanBeDecorated(node as TSNode)) {
       if (ts.isMethodDeclaration(node) && !nodeIsPresent(node.body)) {
-        throw createError(
+        throw new TSError(
           decorator,
           'A decorator can only decorate a method implementation, not an overload.',
         );
       } else {
-        throw createError(decorator, 'Decorators are not valid here.');
+        throw new TSError(decorator, 'Decorators are not valid here.');
       }
     }
   }
@@ -163,7 +163,7 @@ export function checkModifiers(node: ts.Node): void {
         node.kind === SyntaxKind.PropertySignature ||
         node.kind === SyntaxKind.MethodSignature
       ) {
-        throw createError(
+        throw new TSError(
           modifier,
           `'${ts.tokenToString(
             modifier.kind,
@@ -176,7 +176,7 @@ export function checkModifiers(node: ts.Node): void {
         (modifier.kind !== SyntaxKind.StaticKeyword ||
           !ts.isClassLike(node.parent))
       ) {
-        throw createError(
+        throw new TSError(
           modifier,
           `'${ts.tokenToString(
             modifier.kind,
@@ -191,7 +191,7 @@ export function checkModifiers(node: ts.Node): void {
       modifier.kind !== SyntaxKind.ConstKeyword &&
       node.kind === SyntaxKind.TypeParameter
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         `'${ts.tokenToString(
           modifier.kind,
@@ -209,7 +209,7 @@ export function checkModifiers(node: ts.Node): void {
           ts.isTypeAliasDeclaration(node.parent)
         ))
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         `'${ts.tokenToString(
           modifier.kind,
@@ -224,7 +224,7 @@ export function checkModifiers(node: ts.Node): void {
       node.kind !== SyntaxKind.IndexSignature &&
       node.kind !== SyntaxKind.Parameter
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         "'readonly' modifier can only appear on a property declaration or index signature.",
       );
@@ -235,7 +235,7 @@ export function checkModifiers(node: ts.Node): void {
       ts.isClassLike(node.parent) &&
       !ts.isPropertyDeclaration(node)
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         `'${ts.tokenToString(
           modifier.kind,
@@ -249,7 +249,7 @@ export function checkModifiers(node: ts.Node): void {
     ) {
       const declarationKind = getDeclarationKind(node.declarationList);
       if (declarationKind === 'using' || declarationKind === 'await using') {
-        throw createError(
+        throw new TSError(
           modifier,
           `'declare' modifier cannot appear on a '${declarationKind}' declaration.`,
         );
@@ -265,7 +265,7 @@ export function checkModifiers(node: ts.Node): void {
       node.kind !== SyntaxKind.GetAccessor &&
       node.kind !== SyntaxKind.SetAccessor
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         `'${ts.tokenToString(
           modifier.kind,
@@ -281,7 +281,7 @@ export function checkModifiers(node: ts.Node): void {
       (node.parent.kind === SyntaxKind.ModuleBlock ||
         node.parent.kind === SyntaxKind.SourceFile)
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         `'${ts.tokenToString(
           modifier.kind,
@@ -293,7 +293,7 @@ export function checkModifiers(node: ts.Node): void {
       modifier.kind === SyntaxKind.AccessorKeyword &&
       node.kind !== SyntaxKind.PropertyDeclaration
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         "'accessor' modifier can only appear on a property declaration.",
       );
@@ -307,7 +307,7 @@ export function checkModifiers(node: ts.Node): void {
       node.kind !== SyntaxKind.FunctionExpression &&
       node.kind !== SyntaxKind.ArrowFunction
     ) {
-      throw createError(modifier, "'async' modifier cannot be used here.");
+      throw new TSError(modifier, "'async' modifier cannot be used here.");
     }
 
     // `checkGrammarModifiers` function in `typescript`
@@ -318,7 +318,7 @@ export function checkModifiers(node: ts.Node): void {
         modifier.kind === SyntaxKind.DeclareKeyword ||
         modifier.kind === SyntaxKind.AsyncKeyword)
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         `'${ts.tokenToString(
           modifier.kind,
@@ -339,7 +339,7 @@ export function checkModifiers(node: ts.Node): void {
             anotherModifier.kind === SyntaxKind.ProtectedKeyword ||
             anotherModifier.kind === SyntaxKind.PrivateKeyword)
         ) {
-          throw createError(
+          throw new TSError(
             anotherModifier,
             `Accessibility modifier already seen.`,
           );
@@ -363,7 +363,7 @@ export function checkModifiers(node: ts.Node): void {
       if (
         !(func?.kind === SyntaxKind.Constructor && nodeIsPresent(func.body))
       ) {
-        throw createError(
+        throw new TSError(
           modifier,
           'A parameter property is only allowed in a constructor implementation.',
         );
@@ -371,7 +371,7 @@ export function checkModifiers(node: ts.Node): void {
       const param = node as ts.ParameterDeclaration;
 
       if (param.dotDotDotToken) {
-        throw createError(
+        throw new TSError(
           modifier,
           'A parameter property cannot be a rest parameter.',
         );
@@ -381,7 +381,7 @@ export function checkModifiers(node: ts.Node): void {
         param.name.kind === SyntaxKind.ArrayBindingPattern ||
         param.name.kind === SyntaxKind.ObjectBindingPattern
       ) {
-        throw createError(
+        throw new TSError(
           modifier,
           'A parameter property may not be declared using a binding pattern.',
         );
@@ -395,7 +395,7 @@ export function checkModifiers(node: ts.Node): void {
       node.kind === SyntaxKind.MethodDeclaration &&
       node.parent.kind === SyntaxKind.ObjectLiteralExpression
     ) {
-      throw createError(
+      throw new TSError(
         modifier,
         `'${ts.tokenToString(modifier.kind)}' modifier cannot be used here.`,
       );
