+++ packages/typescript-estree/src/check-syntax-errors.ts
@@ -3,9 +3,9 @@ import * as ts from 'typescript';
 import type { TSNode } from './ts-estree';
 
 import { checkModifiers } from './check-modifiers';
+import { TSError } from './errors';
 import {
   isValidAssignmentTarget,
-  createError,
   hasModifier,
   getDeclarationKind,
   getTextForTokenKind,
@@ -31,7 +31,7 @@ export function checkSyntaxError(
           switchCase => switchCase.kind === SyntaxKind.DefaultClause,
         ).length > 1
       ) {
-        throw createError(
+        throw new TSError(
           node,
           "A 'default' clause cannot appear more than once in a 'switch' statement.",
         );
@@ -40,13 +40,13 @@ export function checkSyntaxError(
 
     case SyntaxKind.ThrowStatement:
       if (node.expression.end === node.expression.pos) {
-        throw createError(node, 'A throw statement must throw an expression.');
+        throw new TSError(node, 'A throw statement must throw an expression.');
       }
       break;
 
     case SyntaxKind.CatchClause:
       if (node.variableDeclaration?.initializer) {
-        throw createError(
+        throw new TSError(
           node.variableDeclaration.initializer,
           'Catch clause variable cannot have an initializer.',
         );
@@ -59,23 +59,23 @@ export function checkSyntaxError(
       const isGenerator = !!node.asteriskToken;
       if (isDeclare) {
         if (node.body) {
-          throw createError(
+          throw new TSError(
             node,
             'An implementation cannot be declared in ambient contexts.',
           );
         } else if (isAsync) {
-          throw createError(
+          throw new TSError(
             node,
             "'async' modifier cannot be used in an ambient context.",
           );
         } else if (isGenerator) {
-          throw createError(
+          throw new TSError(
             node,
             'Generators are not allowed in an ambient context.',
           );
         }
       } else if (!node.body && isGenerator) {
-        throw createError(
+        throw new TSError(
           node,
           'A function signature cannot be declared as a generator.',
         );
@@ -88,12 +88,12 @@ export function checkSyntaxError(
 
       if (hasExclamationToken) {
         if (node.initializer) {
-          throw createError(
+          throw new TSError(
             node,
             'Declarations with initializers cannot also have definite assignment assertions.',
           );
         } else if (node.name.kind !== SyntaxKind.Identifier || !node.type) {
-          throw createError(
+          throw new TSError(
             node,
             'Declarations with definite assignment assertions must also have type annotations.',
           );
@@ -108,7 +108,7 @@ export function checkSyntaxError(
           if (
             variableDeclarationList.parent.kind === SyntaxKind.ForInStatement
           ) {
-            throw createError(
+            throw new TSError(
               variableDeclarationList,
               `The left-hand side of a 'for...in' statement cannot be a '${kind}' declaration.`,
             );
@@ -119,13 +119,13 @@ export function checkSyntaxError(
             variableDeclarationList.parent.kind === SyntaxKind.VariableStatement
           ) {
             if (!node.initializer) {
-              throw createError(
+              throw new TSError(
                 node,
                 `'${kind}' declarations must be initialized.`,
               );
             }
             if (node.name.kind !== SyntaxKind.Identifier) {
-              throw createError(
+              throw new TSError(
                 node.name,
                 `'${kind}' declarations may not have binding patterns.`,
               );
@@ -148,7 +148,7 @@ export function checkSyntaxError(
               ['await using', 'const', 'using'].includes(kind)) &&
             hasExclamationToken
           ) {
-            throw createError(
+            throw new TSError(
               node,
               `A definite assignment assertion '!' is not permitted in this context.`,
             );
@@ -159,7 +159,7 @@ export function checkSyntaxError(
             node.initializer &&
             (['let', 'var'].includes(kind) || node.type)
           ) {
-            throw createError(
+            throw new TSError(
               node,
               `Initializers are not permitted in ambient contexts.`,
             );
@@ -182,7 +182,7 @@ export function checkSyntaxError(
       const declarations = node.declarationList.declarations;
 
       if (!declarations.length) {
-        throw createError(
+        throw new TSError(
           node,
           'A variable declaration list must have at least one variable declarator.',
         );
@@ -195,14 +195,14 @@ export function checkSyntaxError(
       const { exclamationToken, questionToken } = node;
 
       if (questionToken) {
-        throw createError(
+        throw new TSError(
           questionToken,
           'A property assignment cannot have a question token.',
         );
       }
 
       if (exclamationToken) {
-        throw createError(
+        throw new TSError(
           exclamationToken,
           'A property assignment cannot have an exclamation token.',
         );
@@ -215,21 +215,21 @@ export function checkSyntaxError(
       const { exclamationToken, modifiers, questionToken } = node;
 
       if (modifiers) {
-        throw createError(
+        throw new TSError(
           modifiers[0],
           'A shorthand property assignment cannot have modifiers.',
         );
       }
 
       if (questionToken) {
-        throw createError(
+        throw new TSError(
           questionToken,
           'A shorthand property assignment cannot have a question token.',
         );
       }
 
       if (exclamationToken) {
-        throw createError(
+        throw new TSError(
           exclamationToken,
           'A shorthand property assignment cannot have an exclamation token.',
         );
@@ -241,7 +241,7 @@ export function checkSyntaxError(
       const isAbstract = hasModifier(SyntaxKind.AbstractKeyword, node);
 
       if (isAbstract && node.initializer) {
-        throw createError(
+        throw new TSError(
           node.initializer,
           `Abstract property cannot have an initializer.`,
         );
@@ -251,7 +251,7 @@ export function checkSyntaxError(
         node.name.kind === SyntaxKind.StringLiteral &&
         node.name.text === 'constructor'
       ) {
-        throw createError(
+        throw new TSError(
           node.name,
           "Classes may not have a field named 'constructor'.",
         );
@@ -261,7 +261,7 @@ export function checkSyntaxError(
 
     case SyntaxKind.TaggedTemplateExpression:
       if (node.tag.flags & ts.NodeFlags.OptionalChain) {
-        throw createError(
+        throw new TSError(
           node,
           'Tagged template expressions are not permitted in an optional chain.',
         );
@@ -273,12 +273,12 @@ export function checkSyntaxError(
         node.operatorToken.kind !== SyntaxKind.InKeyword &&
         node.left.kind === SyntaxKind.PrivateIdentifier
       ) {
-        throw createError(
+        throw new TSError(
           node.left,
           "Private identifiers cannot appear on the right-hand-side of an 'in' expression.",
         );
       } else if (node.right.kind === SyntaxKind.PrivateIdentifier) {
-        throw createError(
+        throw new TSError(
           node.right,
           "Private identifiers are only allowed on the left-hand-side of an 'in' expression.",
         );
@@ -287,7 +287,7 @@ export function checkSyntaxError(
 
     case SyntaxKind.MappedType:
       if (node.members && node.members.length > 0) {
-        throw createError(
+        throw new TSError(
           node.members[0],
           'A mapped type may not declare properties or methods.',
         );
@@ -298,7 +298,7 @@ export function checkSyntaxError(
       // eslint-disable-next-line @typescript-eslint/no-deprecated
       const { initializer } = node;
       if (initializer) {
-        throw createError(
+        throw new TSError(
           initializer,
           'A property signature cannot have an initializer.',
         );
@@ -310,7 +310,7 @@ export function checkSyntaxError(
       // eslint-disable-next-line @typescript-eslint/no-deprecated
       const { modifiers } = node;
       if (modifiers) {
-        throw createError(
+        throw new TSError(
           modifiers[0],
           'A function type cannot have modifiers.',
         );
@@ -321,7 +321,7 @@ export function checkSyntaxError(
     case SyntaxKind.EnumMember: {
       const computed = node.name.kind === ts.SyntaxKind.ComputedPropertyName;
       if (computed) {
-        throw createError(
+        throw new TSError(
           node.name,
           'Computed property names are not allowed in enums.',
         );
@@ -331,7 +331,7 @@ export function checkSyntaxError(
         node.name.kind === SyntaxKind.NumericLiteral ||
         node.name.kind === SyntaxKind.BigIntLiteral
       ) {
-        throw createError(
+        throw new TSError(
           node.name,
           'An enum member cannot have a numeric name.',
         );
@@ -341,7 +341,7 @@ export function checkSyntaxError(
 
     case SyntaxKind.ExternalModuleReference:
       if (node.expression.kind !== SyntaxKind.StringLiteral) {
-        throw createError(node.expression, 'String literal expected.');
+        throw new TSError(node.expression, 'String literal expected.');
       }
       break;
 
@@ -355,7 +355,7 @@ export function checkSyntaxError(
         (operator === '++' || operator === '--') &&
         !isValidAssignmentTarget(node.operand)
       ) {
-        throw createError(
+        throw new TSError(
           node.operand,
           'Invalid left-hand side expression in unary operation',
         );
@@ -373,7 +373,7 @@ export function checkSyntaxError(
         importClause.name &&
         importClause.namedBindings
       ) {
-        throw createError(
+        throw new TSError(
           importClause,
           'A type-only import can specify a default import or named bindings, but not both.',
         );
@@ -398,7 +398,7 @@ export function checkSyntaxError(
         parent.kind === SyntaxKind.ExportDeclaration &&
         parent.moduleSpecifier?.kind !== SyntaxKind.StringLiteral
       ) {
-        throw createError(
+        throw new TSError(
           local,
           'A string literal cannot be used as a local exported binding without `from`.',
         );
@@ -412,7 +412,7 @@ export function checkSyntaxError(
         node.arguments.length !== 1 &&
         node.arguments.length !== 2
       ) {
-        throw createError(
+        throw new TSError(
           node.arguments.length > 1 ? node.arguments[2] : node,
           'Dynamic import requires exactly one or two arguments.',
         );
@@ -425,7 +425,7 @@ export function checkSyntaxError(
         (!hasModifier(ts.SyntaxKind.ExportKeyword, node) ||
           !hasModifier(ts.SyntaxKind.DefaultKeyword, node))
       ) {
-        throw createError(
+        throw new TSError(
           node,
           "A class declaration without the 'default' modifier must have a name.",
         );
@@ -439,7 +439,7 @@ export function checkSyntaxError(
         const { token, types } = heritageClause;
 
         if (types.length === 0) {
-          throw createError(
+          throw new TSError(
             heritageClause,
             `'${ts.tokenToString(token)}' list cannot be empty.`,
           );
@@ -447,18 +447,18 @@ export function checkSyntaxError(
 
         if (token === SyntaxKind.ExtendsKeyword) {
           if (seenExtendsClause) {
-            throw createError(heritageClause, "'extends' clause already seen.");
+            throw new TSError(heritageClause, "'extends' clause already seen.");
           }
 
           if (seenImplementsClause) {
-            throw createError(
+            throw new TSError(
               heritageClause,
               "'extends' clause must precede 'implements' clause.",
             );
           }
 
           if (types.length > 1) {
-            throw createError(
+            throw new TSError(
               types[1],
               'Classes can only extend a single class.',
             );
@@ -468,7 +468,7 @@ export function checkSyntaxError(
         } else {
           // `implements`
           if (seenImplementsClause) {
-            throw createError(
+            throw new TSError(
               heritageClause,
               "'implements' clause already seen.",
             );
@@ -485,7 +485,7 @@ export function checkSyntaxError(
       let seenExtendsClause = false;
       for (const heritageClause of interfaceHeritageClauses) {
         if (heritageClause.token !== SyntaxKind.ExtendsKeyword) {
-          throw createError(
+          throw new TSError(
             heritageClause,
             // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
             heritageClause.token === SyntaxKind.ImplementsKeyword
@@ -494,7 +494,7 @@ export function checkSyntaxError(
           );
         }
         if (seenExtendsClause) {
-          throw createError(heritageClause, "'extends' clause already seen.");
+          throw new TSError(heritageClause, "'extends' clause already seen.");
         }
         seenExtendsClause = true;
 
@@ -503,7 +503,7 @@ export function checkSyntaxError(
             !isEntityNameExpression(heritageType.expression) ||
             ts.isOptionalChain(heritageType.expression)
           ) {
-            throw createError(
+            throw new TSError(
               heritageType,
               'Interface declaration can only extend an identifier/qualified name with optional type arguments.',
             );
@@ -525,7 +525,7 @@ export function checkSyntaxError(
     case SyntaxKind.MethodDeclaration: {
       const isAbstract = hasModifier(SyntaxKind.AbstractKeyword, node);
       if (isAbstract && node.body) {
-        throw createError(
+        throw new TSError(
           node.name,
           node.kind === SyntaxKind.GetAccessor ||
             node.kind === SyntaxKind.SetAccessor
@@ -545,7 +545,7 @@ export function checkSyntaxError(
               property.kind === SyntaxKind.MethodDeclaration) &&
             !property.body
           ) {
-            throw createError(
+            throw new TSError(
               property.end - 1,
               "'{' expected.",
               node.getSourceFile(),
@@ -573,19 +573,19 @@ function checkForStatementDeclaration(
   const loop = kind === SyntaxKind.ForInStatement ? 'for...in' : 'for...of';
   if (ts.isVariableDeclarationList(initializer)) {
     if (initializer.declarations.length !== 1) {
-      throw createError(
+      throw new TSError(
         initializer,
         `Only a single variable declaration is allowed in a '${loop}' statement.`,
       );
     }
     const declaration = initializer.declarations[0];
     if (declaration.initializer) {
-      throw createError(
+      throw new TSError(
         declaration,
         `The variable declaration of a '${loop}' statement cannot have an initializer.`,
       );
     } else if (declaration.type) {
-      throw createError(
+      throw new TSError(
         declaration,
         `The variable declaration of a '${loop}' statement cannot have a type annotation.`,
       );
@@ -595,7 +595,7 @@ function checkForStatementDeclaration(
     initializer.kind !== SyntaxKind.ObjectLiteralExpression &&
     initializer.kind !== SyntaxKind.ArrayLiteralExpression
   ) {
-    throw createError(
+    throw new TSError(
       initializer,
       `The left-hand side of a '${loop}' statement must be a variable or a property access.`,
     );
@@ -607,14 +607,14 @@ function assertModuleSpecifier(
   allowNull: boolean,
 ) {
   if (!allowNull && node.moduleSpecifier == null) {
-    throw createError(node, 'Module specifier must be a string literal.');
+    throw new TSError(node, 'Module specifier must be a string literal.');
   }
 
   if (
     node.moduleSpecifier &&
     node.moduleSpecifier.kind !== SyntaxKind.StringLiteral
   ) {
-    throw createError(
+    throw new TSError(
       node.moduleSpecifier,
       'Module specifier must be a string literal.',
     );
