+++ packages/typescript-estree/src/node-utils.ts
@@ -636,80 +636,6 @@ export function convertTokens(ast: ts.SourceFile): TSESTree.Token[] {
   return result;
 }
 
-export class TSError extends Error {
-  override name = 'TSError';
-
-  constructor(
-    message: string,
-    public readonly fileName: string,
-    public readonly location: {
-      end: {
-        column: number;
-        line: number;
-        offset: number;
-      };
-      start: {
-        column: number;
-        line: number;
-        offset: number;
-      };
-    },
-  ) {
-    super(message);
-  }
-
-  // For old version of ESLint https://github.com/typescript-eslint/typescript-eslint/pull/6556#discussion_r1123237311
-  get index(): number {
-    return this.location.start.offset;
-  }
-
-  // https://github.com/eslint/eslint/blob/b09a512107249a4eb19ef5a37b0bd672266eafdb/lib/linter/linter.js#L853
-  get lineNumber(): number {
-    return this.location.start.line;
-  }
-
-  // https://github.com/eslint/eslint/blob/b09a512107249a4eb19ef5a37b0bd672266eafdb/lib/linter/linter.js#L854
-  get column(): number {
-    return this.location.start.column;
-  }
-}
-
-export function createError(node: ts.Node, message: string): TSError;
-export function createError(
-  node: number | ts.Node | TSESTree.Range,
-  message: string,
-  sourceFile: ts.SourceFile,
-): TSError;
-export function createError(
-  node: number | ts.Node | TSESTree.Range,
-  message: string,
-  sourceFile?: ts.SourceFile,
-): TSError {
-  let startIndex;
-  let endIndex;
-  if (Array.isArray(node)) {
-    [startIndex, endIndex] = node;
-  } else if (typeof node === 'number') {
-    startIndex = endIndex = node;
-  } else {
-    sourceFile ??= node.getSourceFile();
-    startIndex = node.getStart(sourceFile);
-    endIndex = node.getEnd();
-  }
-
-  if (!sourceFile) {
-    throw new Error('`sourceFile` is required.');
-  }
-
-  const [start, end] = [startIndex, endIndex].map(offset => {
-    const { character: column, line } =
-      sourceFile.getLineAndCharacterOfPosition(offset);
-    return { column, line: line + 1, offset };
-  });
-
-  return new TSError(message, sourceFile.fileName, { end, start });
-}
-
 export function nodeHasTokens(n: ts.Node, ast: ts.SourceFile): boolean {
   // If we have a token or node that has a non-zero width, it must have tokens.
   // Note: getWidth() does not take trivia into account.
